yake_demo/requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-debug
flask-wtf
cherrypy
flask
paste
stop-words
flasgger
langdetect
requests
newspaper3k
lxml
Flask-Analytics
git+https://github.com/LIAAD/yake
werkzeug
flask